rename user-config to user_config
